Bug fix and enhancement in Entity classes.

+++ movie-ticket-booking-system/src/main/java/com/movie/entity/Movie.java
@@ -28,7 +28,7 @@ public class Movie {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int movieId;
+	private Long movieId;
 
 	@Column(name = "movie_name", nullable = false, unique = true)
 	private String movieName;
