+++ movie-ticket-booking-system/src/main/java/com/movie/entity/Seat.java
@@ -26,7 +26,7 @@ public class Seat {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int seatId;
+	private Long seatId;
 	private String seatNumber;
 
 	private String seatStatus;
